requirements.txt
flask
pymysql
transformers
torch
python-dotenv
scikit-learn
numpy<2
streamlit